Saved Changes Before Merge to dev

+++ isd-farmers-market/src/main/java/com/example/isdfarmersmarket/web/commands/ItemInCartCommand.java
@@ -8,5 +8,5 @@ import lombok.experimental.FieldDefaults;
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class ItemInCartCommand {
     Long productId;
-    Float quantity;
+    int quantity;
 }
